100% OK until ft_itoa

+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: nrey <marvin@42lausanne.ch>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/03 18:14:07 by nrey              #+#    #+#             */
-/*   Updated: 2024/10/03 18:14:16 by nrey             ###   ########.fr       */
+/*   Updated: 2024/10/05 06:52:52 by nrey             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
